+++ Level0-Module2/src/_01_random/_2_rock_paper_scissors/scissors.jpeg
