+++ Level0-Module2/src/_01_random/_3_dice_roll/dice 1.png
