+++ Level0-Module2/src/_01_random/_4_validation/Validation.java
@@ -1,0 +1,25 @@
+//    Copyright (c) The League of Amazing Programmers 2013-2019
+//    Level 0
+
+package _01_random._4_validation;
+
+import java.util.Random;
+
+import javax.swing.JOptionPane;
+
+public class Validation {
+	public static void main(String[] args) {
+		
+		Random randomMaker = new Random();
+		
+		int randomNumber = randomMaker.nextInt(5);
+		
+		System.out.println(randomNumber);
+
+		// 1. Use each value of randomNumber to give the user a random compliment.
+
+		// 2. Repeat all the code above 10 times
+		
+		// 3. Find someone to test out your program. They will like it :)
+	}
+}
