+++ Level0-Module2/src/_01_random/_6_lottery_numbers/recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Module2/index.html"}
+window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Recipes/LotteryNumbers.html"}
 </script>
