+++ Level0-Module2/src/_02_methods/_6_slippery_dot/slippery_dot/recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Module2/index.html"}
+window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Recipes/SlipperyDot.html"}
 </script>
