+++ Level0-Module2/src/_03_string_conversion/_1_/intro/recipe.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Module2/index.html"}
+window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Recipes/introToParseInt.html"}
 </script>
