--- Level0-Module2/src/_08_safe_cracker/SafeCracker.java
@@ -1,52 +1,0 @@
-//    Copyright (c) The League of Amazing Programmers 2013-2019
-//    Level 0
-
-
-package _08_safe_cracker;
-
-import java.applet.AudioClip;
-import java.io.File;
-import java.net.MalformedURLException;
-import java.util.Random;
-
-import javax.sound.sampled.AudioInputStream;
-import javax.sound.sampled.AudioSystem;
-import javax.sound.sampled.Clip;
-import javax.swing.JApplet;
-import javax.swing.JOptionPane;
-
-public class SafeCracker {
-
-	public static void main(String[] args) {
-		/*
-		 * Your mission: use the tryCode method to crack the safe by trying all possible
-		 * combinations
-		 */
-		
-
-	}
-
-	static void tryCode(int guess) {
-		System.out.println("trying " + guess);
-
-		int secretCode = 9999999 - wekncrzpasfdkjhcfjse;
-
-		if (guess == secretCode) {
-			JOptionPane.showMessageDialog(null, "Congratulations! You cracked the safe with " + guess);
-			playTheSoundOfSuccess();
-			System.exit(0);
-		}
-	}
-
-	private static void playTheSoundOfSuccess() {
-		try {
-			AudioClip sound = JApplet.newAudioClip(SafeCracker.class.getResource("me-gusta.wav"));
-			sound.play();
-			Thread.sleep(3400);
-		} catch (Exception ex) {
-			ex.printStackTrace();
-		}
-	}
-
-	final static int wekncrzpasfdkjhcfjse = new Random().nextInt(100);
-}
